# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_parse_resume_api_resume_parse_post": {
        "properties": {
          "enrich_skills": {
            "default": false,
            "title": "Enrich Skills",
            "type": "boolean"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_parse_resume_api_resume_parse_post",
        "type": "object"
      },
...
      "ContactInfo": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "github": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github"
          },
          "linkedin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "title": "ContactInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResumeListItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "source_available": {
            "default": false,
            "title": "Source Available",
            "type": "boolean"
          },
          "source_format": {
            "anyOf": [
              {
                "enum": [
                  "pdf",
                  "docx"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Format"
          }
        },
        "required": [
          "id",
          "filename",
          "created_at"
        ],
        "title": "ResumeListItem",
        "type": "object"
      },
      "ResumeListResponse": {
        "properties": {
          "resumes": {
            "items": {
              "$ref": "#/components/schemas/ResumeListItem"
            },
            "title": "Resumes",
            "type": "array"
          }
        },
        "required": [
          "resumes"
        ],
        "title": "ResumeListResponse",
        "type": "object"
      },
      "ResumeParseResponse": {
        "properties": {
          "contact_info": {
            "$ref": "#/components/schemas/ContactInfo"
          },
          "enrichment_state": {
            "default": "not_requested",
            "title": "Enrichment State",
            "type": "string"
          },
          "enrichment_units": {
            "default": 0,
            "title": "Enrichment Units",
            "type": "integer"
          },
          "experience_years": {
            "title": "Experience Years",
            "type": "number"
          },
          "extraction_mode": {
            "default": "deterministic",
            "enum": [
              "deterministic",
              "enriched"
            ],
            "title": "Extraction Mode",
            "type": "string"
          },
          "resume_id": {
            "title": "Resume Id",
            "type": "integer"
          },
          "sections": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Sections",
            "type": "object"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "title": "Skills",
            "type": "array"
          },
          "source_available": {
            "default": false,
            "title": "Source Available",
            "type": "boolean"
          },
          "source_format": {
            "anyOf": [
              {
                "enum": [
                  "pdf",
                  "docx"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Format"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "resume_id",
          "skills",
          "experience_years",
          "sections",
          "contact_info"
        ],
        "title": "ResumeParseResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "HireWiz API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/resume/list": {
      "get": {
        "description": "Return all parsed resumes for the current user (used for dropdown selection).",
        "operationId": "list_resumes_api_resume_list_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Resumes",
        "tags": [
          "resume"
        ]
      }
    },
    "/api/resume/parse": {
      "post": {
        "operationId": "parse_resume_api_resume_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_resume_api_resume_parse_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeParseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Parse Resume",
        "tags": [
          "resume"
        ]
      }
    },
    "/api/resume/{resume_id}": {
      "get": {
        "operationId": "get_resume_api_resume__resume_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "title": "Resume Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeParseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Resume",
        "tags": [
          "resume"
        ]
      }
    },
    "/api/resume/{resume_id}/source": {
      "get": {
        "operationId": "download_resume_source_api_resume__resume_id__source_get",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "title": "Resume Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Original uploaded resume"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Download Resume Source",
        "tags": [
          "resume"
        ]
      }
    },
...
  }
}